requirements.txt
flask>=3.0.3
python-dotenv>=1.0.1
authlib>=1.0
requests>=2.27.1
alembic==1.13.2
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
